Keep synced with master

+++ server/src/main/java/com/vaadin/server/VaadinServlet.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2000-2016 Vaadin Ltd.
+ * Copyright 2000-2018 Vaadin Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -15,6 +15,8 @@
  */
 package com.vaadin.server;
 
+import static java.nio.charset.StandardCharsets.UTF_8;
+
 import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileInputStream;
@@ -26,11 +28,13 @@ import java.io.OutputStream;
 import java.io.OutputStreamWriter;
 import java.io.PrintWriter;
 import java.io.Serializable;
+import java.io.UnsupportedEncodingException;
 import java.lang.reflect.Method;
 import java.net.MalformedURLException;
 import java.net.URISyntaxException;
 import java.net.URL;
 import java.net.URLConnection;
+import java.net.URLDecoder;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
@@ -44,7 +48,9 @@ import java.util.Map;
 import java.util.Properties;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+import java.util.regex.Pattern;
 
+import javax.servlet.ServletConfig;
 import javax.servlet.ServletContext;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
@@ -200,32 +206,8 @@ public class VaadinServlet extends HttpServlet implements Constants {
             throws ServletException {
         CurrentInstance.clearAll();
         super.init(servletConfig);
-        Properties initParameters = new Properties();
-
-        readUiFromEnclosingClass(initParameters);
-
-        readConfigurationAnnotation(initParameters);
-
-        // Read default parameters from server.xml
-        final ServletContext context = servletConfig.getServletContext();
-        for (final Enumeration<String> e = context.getInitParameterNames(); e
-                .hasMoreElements();) {
-            final String name = e.nextElement();
-            initParameters.setProperty(name, context.getInitParameter(name));
-        }
-
-        // Override with application config from web.xml
-        for (final Enumeration<String> e = servletConfig
-                .getInitParameterNames(); e.hasMoreElements();) {
-            final String name = e.nextElement();
-            initParameters.setProperty(name,
-                    servletConfig.getInitParameter(name));
-        }
-
-        DeploymentConfiguration deploymentConfiguration = createDeploymentConfiguration(
-                initParameters);
         try {
-            servletService = createServletService(deploymentConfiguration);
+            servletService = createServletService();
         } catch (ServiceException e) {
             throw new ServletException("Could not initialize VaadinServlet", e);
         }
@@ -321,11 +303,96 @@ public class VaadinServlet extends HttpServlet implements Constants {
         }
     }
 
+    /**
+     * Creates a deployment configuration to be used for the creation of a
+     * {@link VaadinService}. Intended to be used by dependency injection
+     * frameworks.
+     *
+     * @return the created deployment configuration
+     *
+     * @throws ServletException
+     *             if construction of the {@link Properties} for
+     *             {@link #createDeploymentConfiguration(Properties)} fails
+     *
+     * @since 8.2
+     */
+    protected DeploymentConfiguration createDeploymentConfiguration()
+            throws ServletException {
+        Properties initParameters = new Properties();
+
+        readUiFromEnclosingClass(initParameters);
+
+        readConfigurationAnnotation(initParameters);
+
+        // Read default parameters from server.xml
+        final ServletContext context = getServletConfig().getServletContext();
+        for (final Enumeration<String> e = context.getInitParameterNames(); e
+                .hasMoreElements();) {
+            final String name = e.nextElement();
+            initParameters.setProperty(name, context.getInitParameter(name));
+        }
+
+        // Override with application config from web.xml
+        for (final Enumeration<String> e = getServletConfig()
+                .getInitParameterNames(); e.hasMoreElements();) {
+            final String name = e.nextElement();
+            initParameters.setProperty(name,
+                    getServletConfig().getInitParameter(name));
+        }
+
+        return createDeploymentConfiguration(initParameters);
+    }
+
+    /**
+     * Creates a deployment configuration to be used for the creation of a
+     * {@link VaadinService}. Override this if you want to override certain
+     * properties.
+     *
+     * @param initParameters
+     *            the context-param and init-param values as properties
+     * @return the created deployment configuration
+     *
+     * @since 7.0.0
+     */
     protected DeploymentConfiguration createDeploymentConfiguration(
             Properties initParameters) {
         return new DefaultDeploymentConfiguration(getClass(), initParameters);
     }
 
+    /**
+     * Creates a vaadin servlet service. This method functions as a layer of
+     * indirection between {@link #init(ServletConfig)} and
+     * {@link #createServletService(DeploymentConfiguration)} so dependency
+     * injection frameworks can call {@link #createDeploymentConfiguration()}
+     * when creating a vaadin servlet service lazily.
+     *
+     * @return the created vaadin servlet service
+     *
+     * @throws ServletException
+     *             if creating a deployment configuration fails
+     * @throws ServiceException
+     *             if creating the vaadin servlet service fails
+     *
+     * @since 8.2
+     */
+    protected VaadinServletService createServletService()
+            throws ServletException, ServiceException {
+        return createServletService(createDeploymentConfiguration());
+    }
+
+    /**
+     * Creates a vaadin servlet service.
+     *
+     * @param deploymentConfiguration
+     *            the deployment configuration to be used
+     *
+     * @return the created vaadin servlet service
+     *
+     * @throws ServiceException
+     *             if creating the vaadin servlet service fails
+     *
+     * @since 7.0.0
+     */
     protected VaadinServletService createServletService(
             DeploymentConfiguration deploymentConfiguration)
             throws ServiceException {
@@ -411,7 +478,7 @@ public class VaadinServlet extends HttpServlet implements Constants {
                 location.length() - lastPathParameter.length());
 
         if ((request.getPathInfo() == null || "/".equals(request.getPathInfo()))
-                && "".equals(request.getServletPath())
+                && request.getServletPath().isEmpty()
                 && !location.endsWith("/")) {
             /*
              * Path info is for the root but request URI doesn't end with a
@@ -471,7 +538,7 @@ public class VaadinServlet extends HttpServlet implements Constants {
     }
 
     /**
-     * Create a Vaadin request for a http servlet request. This method can be
+     * Creates a Vaadin request for a http servlet request. This method can be
      * overridden if the Vaadin request should have special properties.
      *
      * @param request
@@ -514,7 +581,7 @@ public class VaadinServlet extends HttpServlet implements Constants {
                 SystemMessages systemMessages = getService().getSystemMessages(
                         ServletPortletHelper.findLocale(null, null, request),
                         request);
-                getService().writeStringResponse(response,
+                getService().writeUncachedStringResponse(response,
                         JsonConstants.JSON_CONTENT_TYPE,
                         VaadinService.createCriticalNotificationJSON(
                                 systemMessages.getCookiesDisabledCaption(),
@@ -561,7 +628,7 @@ public class VaadinServlet extends HttpServlet implements Constants {
         if (ServletPortletHelper.isUIDLRequest(request)) {
             String output = VaadinService.createCriticalNotificationJSON(
                     caption, message, details, url);
-            getService().writeStringResponse(response,
+            getService().writeUncachedStringResponse(response,
                     JsonConstants.JSON_CONTENT_TYPE, output);
         } else {
             // Create an HTML reponse with the error
@@ -585,7 +652,7 @@ public class VaadinServlet extends HttpServlet implements Constants {
             if (url != null) {
                 output += "</a>";
             }
-            getService().writeStringResponse(response,
+            getService().writeUncachedStringResponse(response,
                     ApplicationConstants.CONTENT_TYPE_TEXT_HTML_UTF_8, output);
         }
     }
@@ -606,7 +673,7 @@ public class VaadinServlet extends HttpServlet implements Constants {
         final OutputStream out = response.getOutputStream();
         try ( // Set the response type
                 PrintWriter outWriter = new PrintWriter(new BufferedWriter(
-                        new OutputStreamWriter(out, "UTF-8")))) {
+                        new OutputStreamWriter(out, UTF_8)))) {
             outWriter.print(output);
             outWriter.flush();
         }
@@ -695,6 +762,13 @@ public class VaadinServlet extends HttpServlet implements Constants {
     private static boolean scssCompileWarWarningEmitted = false;
 
     /**
+     * Pattern for matching request paths that start with /VAADIN/, multiple
+     * slashes allowed on either side.
+     */
+    private static Pattern staticFileRequestPathPatternVaadin = Pattern
+            .compile("^/+VAADIN/.*");
+
+    /**
      * Returns the default theme. Must never return null.
      *
      * @return
@@ -708,13 +782,17 @@ public class VaadinServlet extends HttpServlet implements Constants {
      * resource to the client.
      *
      * @param request
+     *            The request
      * @param response
-     * @return true if a file was served and the request has been handled, false
-     *         otherwise.
+     *            The response
+     * @return {@code true} if a file was served and the request has been
+     *         handled; {@code false} otherwise.
      * @throws IOException
      * @throws ServletException
+     *
+     * @since 8.5
      */
-    private boolean serveStaticResources(HttpServletRequest request,
+    protected boolean serveStaticResources(HttpServletRequest request,
             HttpServletResponse response) throws IOException, ServletException {
 
         String filePath = getStaticFilePath(request);
@@ -732,11 +810,15 @@ public class VaadinServlet extends HttpServlet implements Constants {
      * @param filename
      *            The filename to serve. Should always start with /VAADIN/.
      * @param request
+     *            The request
      * @param response
+     *            The response
      * @throws IOException
      * @throws ServletException
+     *
+     * @since 8.5
      */
-    private void serveStaticResourcesInVAADIN(String filename,
+    protected void serveStaticResourcesInVAADIN(String filename,
             HttpServletRequest request, HttpServletResponse response)
             throws IOException, ServletException {
 
@@ -832,19 +914,15 @@ public class VaadinServlet extends HttpServlet implements Constants {
      * web.xml using resourceCacheTime (defaults to 1 hour).
      *
      * @param filename
+     *            the filename
      * @return cache lifetime for the given filename in seconds
      */
     protected int getCacheTime(String filename) {
-        /*
-         * GWT conventions:
-         *
-         * - files containing .nocache. will not be cached.
-         *
-         * - files containing .cache. will be cached for one year.
-         *
-         * https://developers.google.com/web-toolkit/doc/latest/
-         * DevGuideCompilingAndDebugging#perfect_caching
-         */
+        // GWT conventions:
+        // - files containing .nocache. will not be cached.
+        // - files containing .cache. will be cached for one year.
+        // https://developers.google.com/web-toolkit/doc/latest/DevGuideCompilingAndDebugging#perfect_caching
+
         if (filename.contains(".nocache.")) {
             return 0;
         }
@@ -958,7 +1036,7 @@ public class VaadinServlet extends HttpServlet implements Constants {
     private void streamContent(HttpServletResponse response, InputStream is)
             throws IOException {
         final OutputStream os = response.getOutputStream();
-        final byte buffer[] = new byte[DEFAULT_BUFFER_SIZE];
+        final byte[] buffer = new byte[DEFAULT_BUFFER_SIZE];
         int bytes;
         while ((bytes = is.read(buffer)) >= 0) {
             os.write(buffer, 0, bytes);
@@ -1280,16 +1358,44 @@ public class VaadinServlet extends HttpServlet implements Constants {
      * @since 8.0
      */
     protected String getStaticFilePath(HttpServletRequest request) {
-        String pathInfo = request.getPathInfo();
-        if (pathInfo == null) {
+        if (request.getPathInfo() == null) {
             return null;
+        }
+        String decodedPath = null;
+        String contextPath = null;
+        try {
+            // pathInfo should be already decoded, but some containers do not
+            // decode it, hence we use getRequestURI instead.
+            decodedPath = URLDecoder.decode(request.getRequestURI(),
+                    StandardCharsets.UTF_8.name());
+            contextPath = URLDecoder.decode(request.getContextPath(),
+                    StandardCharsets.UTF_8.name());
+        } catch (UnsupportedEncodingException e) {
+            throw new RuntimeException("An error occurred during decoding URL.",
+                    e);
+        }
+        // Possible context path needs to be removed
+        String filePath = decodedPath.substring(contextPath.length());
+        String servletPath = request.getServletPath();
+        // Possible servlet path needs to be removed
+        if (!servletPath.isEmpty() && !servletPath.equals("/VAADIN")
+                && filePath.startsWith(servletPath)) {
+            filePath = filePath.substring(servletPath.length());
         }
         // Servlet mapped as /* serves at /VAADIN
         // Servlet mapped as /foo/bar/* serves at /foo/bar/VAADIN
-        if (pathInfo.startsWith("/VAADIN/")) {
-            return pathInfo;
+
+        // Matches request paths /VAADIN/*, //VAADIN/* etc.
+        if (staticFileRequestPathPatternVaadin.matcher(filePath).matches()) {
+            // Remove any extra slashes from the beginning,
+            // later occurrences don't interfere
+            while (filePath.startsWith("//")) {
+                filePath = filePath.substring(1);
+            }
+            return filePath;
         }
-        String servletPrefixedPath = request.getServletPath() + pathInfo;
+
+        String servletPrefixedPath = servletPath + filePath;
         // Servlet mapped as /VAADIN/*
         if (servletPrefixedPath.startsWith("/VAADIN/")) {
             return servletPrefixedPath;
@@ -1335,7 +1441,8 @@ public class VaadinServlet extends HttpServlet implements Constants {
                 + request.getServerName()
                 + (request.isSecure() && request.getServerPort() == 443
                         || !request.isSecure() && request.getServerPort() == 80
-                                ? "" : ":" + request.getServerPort())
+                                ? ""
+                                : ":" + request.getServerPort())
                 + request.getRequestURI());
         String servletPath = "";
         if (request
@@ -1351,7 +1458,7 @@ public class VaadinServlet extends HttpServlet implements Constants {
             servletPath = request.getContextPath() + request.getServletPath();
         }
 
-        if (servletPath.length() == 0
+        if (servletPath.isEmpty()
                 || servletPath.charAt(servletPath.length() - 1) != '/') {
             servletPath += "/";
         }
@@ -1367,7 +1474,9 @@ public class VaadinServlet extends HttpServlet implements Constants {
     @Override
     public void destroy() {
         super.destroy();
-        getService().destroy();
+        if (getService() != null) {
+            getService().destroy();
+        }
     }
 
     private static void persistCacheEntry(ScssCacheEntry cacheEntry) {
@@ -1449,7 +1558,7 @@ public class VaadinServlet extends HttpServlet implements Constants {
             } else {
                 safe.append("&#");
                 safe.append((int) c);
-                safe.append(";");
+                safe.append(';');
             }
         }
